changes to crud and added mysql database

+++ src/main/java/com/biblioteczka/S/BLibrary/Repository/BookRepository.java
@@ -3,6 +3,6 @@ package com.biblioteczka.S.BLibrary.Repository;
 import com.biblioteczka.S.BLibrary.Model.BookModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<BookModel, Long> {
+public interface BookRepository extends JpaRepository<BookModel, Integer> {
 
 }
